Improve Navbar unit tests and
- Also add snapshot validation
